requirements.txt
flask
google-generativeai
pandas
openpyxl
python-dotenv
werkzeug
